refactor: Kafka 프로듀서 설정 추가 및 컨슈머 이름/패키지 변경
